// rpc/generated/nodes.ts
export interface Node {
  node?:
    | { $case: "include"; include: IncludeNode }
    | { $case: "funcDecl"; funcDecl: FunctionDeclarationNode }
    | { $case: "argument"; argument: ArgumentNode }
    | { $case: "variable"; variable: VariableNode }
    | { $case: "literal"; literal: LiteralNode }
    | { $case: "return"; return: ReturnNode }
    | { $case: "funcCall"; funcCall: FunctionCallNode }
    | { $case: "compStmt"; compStmt: CompoundStatementNode }
    | undefined;
}

export interface CompoundStatementNode {
  elements: Node[];
}

export interface IncludeNode {
  id: string;
  directive: string;
}

export interface FunctionDeclarationNode {
  id: string;
  name: string;
  returnType: string;
  arguments: ArgumentNode[];
  body?: CompoundStatementNode | undefined;
}

export interface ArgumentNode {
  id: string;
  name: string;
  dataType: string;
}

export interface VariableNode {
  id: string;
  name: string;
  dataType: string;
  /** Deberia ser un expressionNode */
  initializer?: Node | undefined;
}

export interface LiteralNode {
  value: string;
}

export interface ReturnNode {
  id: string;
  /** Deberia ser un expressionNode */
  expression?: Node | undefined;
}

export interface FunctionCallNode {
  id: string;
  name: string;
  parameters: string[];
}
// ...
export const IncludeNode: MessageFns<IncludeNode> = {
  encode(message: IncludeNode, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.directive !== "") {
      writer.uint32(18).string(message.directive);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IncludeNode {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIncludeNode();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.directive = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): IncludeNode {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      directive: isSet(object.directive) ? globalThis.String(object.directive) : "",
    };
  },

  toJSON(message: IncludeNode): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.directive !== "") {
      obj.directive = message.directive;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<IncludeNode>, I>>(base?: I): IncludeNode {
    return IncludeNode.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<IncludeNode>, I>>(object: I): IncludeNode {
    const message = createBaseIncludeNode();
    message.id = object.id ?? "";
    message.directive = object.directive ?? "";
    return message;
  },
};

function createBaseFunctionDeclarationNode(): FunctionDeclarationNode {
  return { id: "", name: "", returnType: "", arguments: [], body: undefined };
}

export const FunctionDeclarationNode: MessageFns<FunctionDeclarationNode> = {
  encode(message: FunctionDeclarationNode, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.returnType !== "") {
      writer.uint32(26).string(message.returnType);
    }
    for (const v of message.arguments) {
      ArgumentNode.encode(v!, writer.uint32(34).fork()).join();
    }
    if (message.body !== undefined) {
      CompoundStatementNode.encode(message.body, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FunctionDeclarationNode {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFunctionDeclarationNode();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.returnType = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.arguments.push(ArgumentNode.decode(reader, reader.uint32()));
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.body = CompoundStatementNode.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FunctionDeclarationNode {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      returnType: isSet(object.returnType) ? globalThis.String(object.returnType) : "",
      arguments: globalThis.Array.isArray(object?.arguments)
        ? object.arguments.map((e: any) => ArgumentNode.fromJSON(e))
        : [],
      body: isSet(object.body) ? CompoundStatementNode.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: FunctionDeclarationNode): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.returnType !== "") {
      obj.returnType = message.returnType;
    }
    if (message.arguments?.length) {
      obj.arguments = message.arguments.map((e) => ArgumentNode.toJSON(e));
    }
    if (message.body !== undefined) {
      obj.body = CompoundStatementNode.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FunctionDeclarationNode>, I>>(base?: I): FunctionDeclarationNode {
    return FunctionDeclarationNode.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FunctionDeclarationNode>, I>>(object: I): FunctionDeclarationNode {
    const message = createBaseFunctionDeclarationNode();
    message.id = object.id ?? "";
    message.name = object.name ?? "";
    message.returnType = object.returnType ?? "";
    message.arguments = object.arguments?.map((e) => ArgumentNode.fromPartial(e)) || [];
    message.body = (object.body !== undefined && object.body !== null)
      ? CompoundStatementNode.fromPartial(object.body)
      : undefined;
    return message;
  },
};
// ...
export const VariableNode: MessageFns<VariableNode> = {
  encode(message: VariableNode, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.dataType !== "") {
      writer.uint32(26).string(message.dataType);
    }
    if (message.initializer !== undefined) {
      Node.encode(message.initializer, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): VariableNode {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseVariableNode();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.dataType = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.initializer = Node.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): VariableNode {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      dataType: isSet(object.dataType) ? globalThis.String(object.dataType) : "",
      initializer: isSet(object.initializer) ? Node.fromJSON(object.initializer) : undefined,
    };
  },

  toJSON(message: VariableNode): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.dataType !== "") {
      obj.dataType = message.dataType;
    }
    if (message.initializer !== undefined) {
      obj.initializer = Node.toJSON(message.initializer);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<VariableNode>, I>>(base?: I): VariableNode {
    return VariableNode.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<VariableNode>, I>>(object: I): VariableNode {
    const message = createBaseVariableNode();
    message.id = object.id ?? "";
    message.name = object.name ?? "";
    message.dataType = object.dataType ?? "";
    message.initializer = (object.initializer !== undefined && object.initializer !== null)
      ? Node.fromPartial(object.initializer)
      : undefined;
    return message;
  },
};

function createBaseLiteralNode(): LiteralNode {
  return { value: "" };
}

export const LiteralNode: MessageFns<LiteralNode> = {
  encode(message: LiteralNode, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.value !== "") {
      writer.uint32(10).string(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiteralNode {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiteralNode();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.value = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiteralNode {
    return { value: isSet(object.value) ? globalThis.String(object.value) : "" };
  },

  toJSON(message: LiteralNode): unknown {
    const obj: any = {};
    if (message.value !== "") {
      obj.value = message.value;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiteralNode>, I>>(base?: I): LiteralNode {
    return LiteralNode.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiteralNode>, I>>(object: I): LiteralNode {
    const message = createBaseLiteralNode();
    message.value = object.value ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };